Arbitrary hacking to make MacFusion more stable. I'm more seasoned at iOS dev than Mac development and even there I'm a rookie. I'm not exactly sure why I managed to create a build that works correctly on 10.8 but I'll take it.
I have left in some NSLogs for the sake of easy debugging when things go bad.
The problem with the prepackaged binary on the main MacFusion page appeares
to be that the NSAlert callback for deleting an item from the server list
passes a reference that has been garbage collected already. Specifically,
the context being passed to the callback is the NSArray of servers in the list.

Compiling with 10.7's BaseSDK seems to have fixed things. I have no idea why.